+++ pyproject.toml
@@ -201,7 +201,7 @@ module = [
     "neptune.attributes.utils",
     "neptune.common.backends.utils",
     "neptune.common.exceptions",
-    "neptune.common.git_info",
+    "neptune.internal.utils.git_info",
     "neptune.common.hardware.cgroup.cgroup_filesystem_reader",
     "neptune.common.hardware.cgroup.cgroup_monitor",
     "neptune.common.hardware.gauges.cpu",
@@ -221,11 +221,11 @@ module = [
     "neptune.common.hardware.resources.system_resource_info",
     "neptune.common.hardware.resources.system_resource_info_factory",
     "neptune.common.hardware.system.system_monitor",
-    "neptune.common.oauth",
+    "neptune.internal.oauth",
     "neptune.common.patches.bravado",
     "neptune.common.storage.datastream",
     "neptune.common.storage.storage_utils",
-    "neptune.common.utils",
+    "neptune.internal.utils.utils",
     "neptune.internal.warnings",
     "neptune.common.websockets.reconnecting_websocket",
     "neptune.common.websockets.websocket_client_adapter",
