+++ src/neptune/internal/credentials.py
@@ -24,10 +24,10 @@ from typing import (
     Optional,
 )
 
-from neptune.common.envs import API_TOKEN_ENV_NAME
 from neptune.constants import ANONYMOUS_API_TOKEN
 from neptune.exceptions import NeptuneMissingApiTokenException
 from neptune.internal.constants import ANONYMOUS_API_TOKEN_CONTENT
+from neptune.internal.envs import API_TOKEN_ENV_NAME
 from neptune.internal.exceptions import NeptuneInvalidApiTokenException
