+++ src/neptune/internal/utils/git_info.py
