+++ src/neptune/internal/utils/patterns.py
