+++ src/neptune/internal/utils/utils.py
@@ -25,9 +25,7 @@ import sys
 import numpy as np
 import pandas as pd
 
-from neptune.common import envs
-from neptune.common.git_info import GitInfo
-from neptune.common.patterns import PROJECT_QUALIFIED_NAME_PATTERN
+from neptune.internal import envs
 from neptune.internal.exceptions import (
     FileNotFound,
     InvalidNotebookPath,
@@ -36,7 +34,9 @@ from neptune.internal.exceptions import (
     NotADirectory,
     NotAFile,
 )
+from neptune.internal.utils.git_info import GitInfo
 from neptune.internal.utils.logger import get_logger
+from neptune.internal.utils.patterns import PROJECT_QUALIFIED_NAME_PATTERN
 
 _logger = get_logger()
